Added a pet shop class, and menus
There are now 2 menus, one main menu and one menu for selecting the species of a new pet to add. There is also a pet shop class, which stores methods to display and collect menus and menu selections to the user, and will store a list of all of the pets available in the shop.